Refactor: CrawlerService에서 MySQL 저장 과정 삭제, StoreBatchController, TestController 컨트롤러 이름 변경
