+++ src/main/java/com/sangchu/batch/patch/controller/MysqlBatchController.java
@@ -1,10 +1,14 @@
-package com.sangchu.batch.preprocess.controller;
+package com.sangchu.batch.patch.controller;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.batch.core.Job;
 import org.springframework.batch.core.JobParameters;
 import org.springframework.batch.core.JobParametersBuilder;
+import org.springframework.batch.core.JobParametersInvalidException;
 import org.springframework.batch.core.launch.JobLauncher;
+import org.springframework.batch.core.repository.JobExecutionAlreadyRunningException;
+import org.springframework.batch.core.repository.JobInstanceAlreadyCompleteException;
+import org.springframework.batch.core.repository.JobRestartException;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
@@ -12,19 +16,19 @@ import org.springframework.web.bind.annotation.RestController;
 @RestController
 @RequestMapping("/test")
 @RequiredArgsConstructor
-public class StoreBatchController {
+public class MysqlBatchController {
 
     private final JobLauncher jobLauncher;
-    private final Job elasticsearchJob;
+    private final Job mysqlJob;
 
-    @PostMapping("/import/elasticsearch")
-    public String runImportElasticsearchJob() throws Exception {
+    @PostMapping("/import/mysql")
+    public String runImportMysqlJob() throws JobInstanceAlreadyCompleteException, JobExecutionAlreadyRunningException, JobParametersInvalidException, JobRestartException {
         JobParameters jobParameters = new JobParametersBuilder()
                 .addLong("time", System.currentTimeMillis())
                 .toJobParameters();
 
-        jobLauncher.run(elasticsearchJob, jobParameters);
+        jobLauncher.run(mysqlJob, jobParameters);
 
-        return "mysql -> elasticsearch 작업";
+        return "csv -> mysql 작업";
     }
-} 
+}
