+++ src/main/java/com/sangchu/batch/preprocess/controller/ElasticsearchBatchController.java
@@ -1,14 +1,10 @@
-package com.sangchu.batch.patch.controller;
+package com.sangchu.batch.preprocess.controller;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.batch.core.Job;
 import org.springframework.batch.core.JobParameters;
 import org.springframework.batch.core.JobParametersBuilder;
-import org.springframework.batch.core.JobParametersInvalidException;
 import org.springframework.batch.core.launch.JobLauncher;
-import org.springframework.batch.core.repository.JobExecutionAlreadyRunningException;
-import org.springframework.batch.core.repository.JobInstanceAlreadyCompleteException;
-import org.springframework.batch.core.repository.JobRestartException;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
@@ -16,19 +12,19 @@ import org.springframework.web.bind.annotation.RestController;
 @RestController
 @RequestMapping("/test")
 @RequiredArgsConstructor
-public class TestController {
+public class ElasticsearchBatchController {
 
     private final JobLauncher jobLauncher;
-    private final Job mysqlJob;
+    private final Job elasticsearchJob;
 
-    @PostMapping("/import/mysql")
-    public String runImportMysqlJob() throws JobInstanceAlreadyCompleteException, JobExecutionAlreadyRunningException, JobParametersInvalidException, JobRestartException {
+    @PostMapping("/import/elasticsearch")
+    public String runImportElasticsearchJob() throws Exception {
         JobParameters jobParameters = new JobParametersBuilder()
                 .addLong("time", System.currentTimeMillis())
                 .toJobParameters();
 
-        jobLauncher.run(mysqlJob, jobParameters);
+        jobLauncher.run(elasticsearchJob, jobParameters);
 
-        return "csv -> mysql 작업";
+        return "mysql -> elasticsearch 작업";
     }
-}
+}
